fix project structure
