+++ jubapi/controllers/v1/products.py
@@ -1,16 +1,14 @@
 import os
 import time as T
 from typing import List
-from fastapi import APIRouter,Depends,Response,HTTPException
-from fastapi.encoders import jsonable_encoder
-from fastapi.responses import JSONResponse
-from jubapi.repositories.products import ProductsRepository
-from jubapi.repositories.observatory import ObservatoriesRepository
-from jubapi.repositories.catalog import CatalogsRepository
+from fastapi import APIRouter,Depends,Response
+from jubapi.repositories.v1.products import ProductsRepository
+from jubapi.repositories.v1.observatory import ObservatoriesRepository
+from jubapi.repositories.v1.catalog import CatalogsRepository
 from jubapi.db import get_collection
-from jubapi.services import ProductsService,ObservatoriesService,CatalogsService
-from jubapi.dto.product import ProductDTO
-from jubapi.dto import ProductFilter
+from jubapi.services.v1 import ProductsService,ObservatoriesService,CatalogsService
+from jubapi.dto.v1.product import ProductDTO
+from jubapi.dto.v1 import ProductFilter
 from jubapi.log.log import Log
 import jubapi.config as CX
 # LOG_DEBUG = bool(int(os.environ.get("LOG_DEBUG","1")))
