+++ jubapi/repositories/v1/products.py
@@ -1,4 +1,3 @@
-# from pymongo.collection import Collection
 from motor.motor_asyncio import AsyncIOMotorCollection
 from pymongo.results import DeleteResult
 from bson import ObjectId
@@ -6,7 +5,6 @@ from option import Option, NONE, Some,Result,Ok,Err
 from typing import List,Any,Dict
 from jubapi.models.v1 import Product
 from jubapi.dto.v1.product import ProductDTO
-# from jubapi.utils.utils import Utils as U
